Print a warning when more than 512MiB of memory are available
ServerJars doesn't need that much memory, so it is wise to limit its consumtion. We recommend 128M for now but only warn on 512 or above. There might be reasons to have it higher than 128M (special server mods?) but more than 512M are probably due to not setting a limit at all


Related to #23